+++ libs/cardano-ledger-core/src/Cardano/Ledger/BaseTypes.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveAnyClass #-}
@@ -26,6 +27,7 @@ module Cardano.Ledger.BaseTypes (
   module Slotting,
   module NonZero,
   ProtVer (..),
+  decodeProtVer,
   module Cardano.Ledger.Binary.Version,
   FixedPoint,
   (==>),
@@ -111,6 +113,7 @@ import Cardano.Ledger.Binary (
   FromCBOR,
   ToCBOR,
   cborError,
+  decodeIntegralRational,
   encodeListLen,
   fromPlainDecoder,
   ifDecoderVersionAtLeast,
@@ -130,6 +133,7 @@ import Cardano.Ledger.Binary.Plain (
  )
 import qualified Cardano.Ledger.Binary.Plain as Plain
 import Cardano.Ledger.Binary.Version
+import Cardano.Ledger.Core.Era (Era (..))
 import Cardano.Ledger.Hashes (HashAnnotated, SafeHash, SafeToHash)
 import Cardano.Ledger.Keys (KeyHash, KeyRole (..))
 import Cardano.Ledger.NonIntegral (ln')
@@ -239,6 +243,25 @@ instance DecCBORGroup ProtVer where
         (natVersion @12)
         (fromIntegral @Word32 @Natural <$> decCBOR @Word32)
         (decCBOR @Natural)
+
+-- | Decoder for `ProtVer` that only accepts versions where the major fits in
+-- the bounds of what's allowed for the provided Era.
+--
+-- /Note/ - Thanks to the Ledger rules, only the current major protocol version
+-- and the very next major protocol version from the current one will be accepted
+-- by the Ledger anywhere in the block, regardless that this decoder allows a
+-- broader range.
+decodeProtVer :: forall era s. Era era => Decoder s ProtVer
+decodeProtVer = do
+  maxMajorVersion <- succVersion (natVersion @(ProtVerHigh era))
+  pv@(ProtVer major _) <- decCBOR
+  when (major > maxMajorVersion) $
+    fail $
+      "Protocol version "
+        <> show major
+        <> " exceeds the maximum expected version of "
+        <> show maxMajorVersion
+  pure pv
 
 data E34
 
@@ -382,11 +405,27 @@ instance
 instance (ToCBOR (BoundedRatio b a), Typeable b, Typeable a) => EncCBOR (BoundedRatio b a)
 
 instance
-  (FromCBOR (BoundedRatio b a), Typeable b, Typeable a, Typeable (BoundedRatio b a)) =>
+  ( FromCBOR (BoundedRatio b a)
+  , Bounded (BoundedRatio b a)
+  , Bounded a
+  , Integral a
+  , DecCBOR a
+  , Typeable b
+  , Typeable a
+  , Typeable (BoundedRatio b a)
+  ) =>
   DecCBOR (BoundedRatio b a)
   where
-  decCBOR = fromPlainDecoder fromCBOR
-  {-# INLINE decCBOR #-}
+  decCBOR =
+    ifDecoderVersionAtLeast
+      (natVersion @12)
+      ( do
+          r <- decodeIntegralRational @a
+          case boundRational r of
+            Nothing -> cborError $ DecoderErrorCustom "BoundedRatio" (Text.pack $ show r)
+            Just u -> pure u
+      )
+      (fromPlainDecoder fromCBOR)
 
 instance Bounded (BoundedRatio b Word64) => ToJSON (BoundedRatio b Word64) where
   toJSON :: BoundedRatio b Word64 -> Value